Re-org `convert_to_f64_array` to utils
