+++ datafusion/functions-nested/src/utils.rs
@@ -22,16 +22,18 @@ use std::sync::Arc;
 use arrow::{array::ArrayRef, datatypes::DataType};
 
 use arrow_array::{
-    Array, BooleanArray, GenericListArray, ListArray, OffsetSizeTrait, Scalar,
-    UInt32Array,
+    Array, BooleanArray, Float64Array, GenericListArray, ListArray, OffsetSizeTrait,
+    Scalar, UInt32Array,
 };
 use arrow_buffer::OffsetBuffer;
 use arrow_schema::{Field, Fields};
-use datafusion_common::cast::{as_large_list_array, as_list_array};
-use datafusion_common::{exec_err, internal_err, plan_err, Result, ScalarValue};
-
 use core::any::type_name;
+use datafusion_common::cast::{
+    as_float32_array, as_float64_array, as_int32_array, as_int64_array,
+    as_large_list_array, as_list_array,
+};
 use datafusion_common::DataFusionError;
+use datafusion_common::{exec_err, internal_err, plan_err, Result, ScalarValue};
 use datafusion_expr::{ColumnarValue, ScalarFunctionImplementation};
 
 macro_rules! downcast_arg {
@@ -265,6 +267,32 @@ pub(crate) fn get_map_entry_field(data_type: &DataType) -> Result<&Fields> {
             }
         }
         _ => internal_err!("Expected a Map type, got {:?}", data_type),
+    }
+}
+
+/// Converts an array of any numeric type to a Float64Array.
+pub(crate) fn convert_to_f64_array(array: &ArrayRef) -> Result<Float64Array> {
+    match array.data_type() {
+        DataType::Float64 => Ok(as_float64_array(array)?.clone()),
+        DataType::Float32 => {
+            let array = as_float32_array(array)?;
+            let converted: Float64Array =
+                array.iter().map(|v| v.map(|v| v as f64)).collect();
+            Ok(converted)
+        }
+        DataType::Int64 => {
+            let array = as_int64_array(array)?;
+            let converted: Float64Array =
+                array.iter().map(|v| v.map(|v| v as f64)).collect();
+            Ok(converted)
+        }
+        DataType::Int32 => {
+            let array = as_int32_array(array)?;
+            let converted: Float64Array =
+                array.iter().map(|v| v.map(|v| v as f64)).collect();
+            Ok(converted)
+        }
+        _ => exec_err!("Unsupported array type for conversion to Float64Array"),
     }
 }
